+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tutorial on how to load models dynamically
   https://intel.github.io/dffml/tutorials/models/load.html
 - Added download progressbar in `util/net.py`
+- Usecase example notebook for "Moving between models"
+- Documentation and testing support for notebooks
 ### Changed
 - Calls to hashlib now go through helper functions
 - Build docs using `dffml service dev docs`